Extract indexer assignment samples
